Contracts: Update get-bytecode script
